feat: Error message handling while creating or deleting the bucket
